requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.2.0
pyjwt==2.10.1
python-jose==3.3.0